Add files via upload
Data Tools:
As long as you’re able to recognize data patterns, find overlap between data listed, and properly read the data sets, you’ll be able to properly analyze the data to try finding trends that may relate to each other. However, correlation is not causation and just because they share similarities or seem related, does not mean that they are.

Big Data:
When working with computers, not all data is created equally and with more data being generated than ever, it’s important to understand what types of data require more memory and how that’ll affect our computer systems. With newer techniques like parallel computing, which splits up the work into smaller subsets to make the processing more efficient and faster, we can also sort through more data than ever before. However, as we gather more and more data, scalability becomes a major concern and the cornerstone of ensuring that we have the necessary space for our storage to grow. In short, while the challenges of Big Data are titanic, we are able to handle it with new methods of computing and programming.

Bias in machine learning:
In the world we currently have, it’s very easy for our biases to bleed into our work without us even knowing. This is also true for the programs we create. Because a lot of programs rely on external data sets, the bias in those data sets can easily bleed into the program naturally without the programmer even noticing. As a result, it is best for us to rigorously test our programs and to find ways to lower or limit this bias as much as we possibly can.

Unit test:
The unit test was nothing new, actually using questions that were previously asked in the prior sections. The test was straightforward and not particularly challenging but good for anyone unfamiliar with how to analyze data sets.


Hypothesis:
After the end of 1981, we’ll see a dramatic rise in house prices that will even out to an average far above the prices at 1975, but we’ll see a similar surge a few years before the 2008 Housing Crash, leading to a short dip before it begins to rise exponentially again.


Process:
This information was found on catalog.data.gov, which published a comprehensive index on the Housing Prices recorded by the Federal Housing Finance Agency. After transferring this data into Google Sheets, I sorted it by year and made a graph to track the HPI values by year. No need for averages when we can see the data in full in this case. After creating the graph, it was easy to see that the hypothesis I had was correct, as housing prices now soar while we enter into Housing Bubble 2.0.